python/requirements.txt
spacy
psycopg[binary]
python-dotenv
